reflectai-backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
python-dotenv
requests

opencv-python-headless
deepface
numpy

openai-whisper
google-generativeai

